+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See `UPGRADE.md` for instructions on upgrading between versions.
 * Live preview: Changed syntax - see UPGRADE.md
 * Live preview: Now sends entry diffs on update to save some bandwidth
 * Live preview: Send base64 of images on entry creation
+* Villain: Added telemetry for `parse_and_render`
 
 
 ## 0.48.0